App now stores inventory changes, borrow inventory changes and updates the list view between uses. Just needs the robot communication added in

+++ src/main/java/edu/wpi/rbe/rbe2001/fieldsimulator/gui/WelcomeScreenController.java
@@ -54,10 +54,10 @@ public class WelcomeScreenController implements Initializable {
     }
 
     public void enterCallback(){
-        int ID = 0;
+        long ID = 0;
         boolean idValid = true;
         try{
-            ID = Integer.parseInt(IDEntry.getText());
+            ID = Long.parseLong(IDEntry.getText());
             if(IDEntry.getText().length()!=9){
                 idValid = false;
             }
